+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Stencil Changelog
 
+## Master
+
+### Bug Fixes
+
+- Fixed using quote as a filter parameter
+
+
 ## 0.11.0 (2018-04-04)
 
 ### Enhancements